Apply forge fmt.

+++ src/Voter.sol
@@ -21,11 +21,7 @@ import "./interfaces/dex/IPearlV2Pool.sol";
 import "./interfaces/dex/IPearlV2Factory.sol";
 import "./Epoch.sol";
 
-contract Voter is
-    IVoter,
-    NonblockingLzAppUpgradeable,
-    ReentrancyGuardUpgradeable
-{
+contract Voter is IVoter, NonblockingLzAppUpgradeable, ReentrancyGuardUpgradeable {
     /// @custom:oz-upgrades-unsafe-allow state-variable-immutable
     bool public immutable isMainChain;
 
@@ -84,36 +80,15 @@ contract Voter is
     event GaugeRevived(address indexed gauge);
     event Voted(address indexed voter, uint256 weight);
     event Abstained(address indexed voter, uint256 weight);
-    event Deposit(
-        address indexed lp,
-        address indexed gauge,
-        uint256 tokenId,
-        uint256 amount
-    );
-    event Withdraw(
-        address indexed lp,
-        address indexed gauge,
-        uint256 tokenId,
-        uint256 amount
-    );
-    event NotifyReward(
-        address indexed sender,
-        address indexed reward,
-        uint256 amount
-    );
-    event DistributeReward(
-        address indexed sender,
-        address indexed gauge,
-        uint256 amount
-    );
+    event Deposit(address indexed lp, address indexed gauge, uint256 tokenId, uint256 amount);
+    event Withdraw(address indexed lp, address indexed gauge, uint256 tokenId, uint256 amount);
+    event NotifyReward(address indexed sender, address indexed reward, uint256 amount);
+    event DistributeReward(address indexed sender, address indexed gauge, uint256 amount);
     event Whitelisted(address indexed whitelister, address indexed token);
     event LzAdapterParams(uint256 limit);
 
     /// @custom:oz-upgrades-unsafe-allow constructor
-    constructor(
-        uint256 _mainChainId,
-        address _lzEndpoint
-    ) NonblockingLzAppUpgradeable(_lzEndpoint) {
+    constructor(uint256 _mainChainId, address _lzEndpoint) NonblockingLzAppUpgradeable(_lzEndpoint) {
         isMainChain = _mainChainId == block.chainid;
         _disableInitializers();
     }
@@ -130,12 +105,8 @@ contract Voter is
         uint16 _lzChainId
     ) public initializer {
         require(
-            _initialOwner != address(0) &&
-                _lockedToken != address(0) &&
-                _emergencyCouncil != address(0) &&
-                _factory != address(0) &&
-                _gauges != address(0) &&
-                _bribes != address(0),
+            _initialOwner != address(0) && _lockedToken != address(0) && _emergencyCouncil != address(0)
+                && _factory != address(0) && _gauges != address(0) && _bribes != address(0),
             "!zero address"
         );
 
@@ -151,10 +122,7 @@ contract Voter is
         // VotingEscrow is only deployed on main chain
         if (__ve != address(0)) {
             _ve = __ve;
-            require(
-                _lockedToken == address(IVotingEscrow(__ve).lockedToken()),
-                "!locked token"
-            );
+            require(_lockedToken == address(IVotingEscrow(__ve).lockedToken()), "!locked token");
         }
 
         factory = _factory;
@@ -190,10 +158,7 @@ contract Voter is
 
     function _checkDistribution() internal view {
         require(epochController != address(0), "!epochController");
-        require(
-            !IEpochController(epochController).checkDistribution(),
-            "Voting is not allowed during distribution"
-        );
+        require(!IEpochController(epochController).checkDistribution(), "Voting is not allowed during distribution");
     }
 
     function _updateFor(address _gauge) internal {
@@ -228,14 +193,8 @@ contract Voter is
                 weights[_pool] -= _votes;
                 votes[account][_pool] -= _votes;
                 if (_votes > 0) {
-                    IBribe(internal_bribes[gauges[_pool]])._withdraw(
-                        uint256(_votes),
-                        account
-                    );
-                    IBribe(external_bribes[gauges[_pool]])._withdraw(
-                        uint256(_votes),
-                        account
-                    );
+                    IBribe(internal_bribes[gauges[_pool]])._withdraw(uint256(_votes), account);
+                    IBribe(external_bribes[gauges[_pool]])._withdraw(uint256(_votes), account);
                     _totalWeight += _votes;
                 } else {
                     _totalWeight -= _votes;
@@ -252,7 +211,7 @@ contract Voter is
         address[] memory _poolVote = poolVote[account];
         uint256 _poolCnt = _poolVote.length;
         uint256[] memory _weights = new uint256[](_poolCnt);
-        for (uint256 i = _poolCnt; i != 0; ) {
+        for (uint256 i = _poolCnt; i != 0;) {
             unchecked {
                 --i;
             }
@@ -261,25 +220,21 @@ contract Voter is
         _vote(account, _poolVote, _weights);
     }
 
-    function _vote(
-        address account,
-        address[] memory _poolVote,
-        uint256[] memory _weights
-    ) internal {
+    function _vote(address account, address[] memory _poolVote, uint256[] memory _weights) internal {
         _reset(account);
         uint256 _poolCnt = _poolVote.length;
         uint256 _weight = IVotingEscrow(_ve).getVotes(account);
         uint256 _totalVoteWeight = 0;
         uint256 _usedWeight = 0;
 
-        for (uint256 i = _poolCnt; i != 0; ) {
+        for (uint256 i = _poolCnt; i != 0;) {
             unchecked {
                 --i;
             }
             _totalVoteWeight += _weights[i];
         }
 
-        for (uint256 i = _poolCnt; i != 0; ) {
+        for (uint256 i = _poolCnt; i != 0;) {
             unchecked {
                 --i;
             }
@@ -287,8 +242,7 @@ contract Voter is
             address _gauge = gauges[_pool];
 
             if (isGauge[_gauge]) {
-                uint256 _poolWeight = (_weights[i] * _weight) /
-                    _totalVoteWeight;
+                uint256 _poolWeight = (_weights[i] * _weight) / _totalVoteWeight;
                 require(votes[account][_pool] == 0);
                 require(_poolWeight != 0);
 
@@ -297,15 +251,9 @@ contract Voter is
                 weights[_pool] += _poolWeight;
                 votes[account][_pool] += _poolWeight;
 
-                IBribe(internal_bribes[_gauge])._deposit(
-                    uint256(_poolWeight),
-                    account
-                );
+                IBribe(internal_bribes[_gauge])._deposit(uint256(_poolWeight), account);
 
-                IBribe(external_bribes[_gauge])._deposit(
-                    uint256(_poolWeight),
-                    account
-                );
+                IBribe(external_bribes[_gauge])._deposit(uint256(_poolWeight), account);
                 _usedWeight += _poolWeight;
                 emit Voted(account, _poolWeight);
             }
@@ -315,22 +263,11 @@ contract Voter is
         usedWeights[account] = _usedWeight;
     }
 
-    function _safeTransferFrom(
-        address token,
-        address from,
-        address to,
-        uint256 value
-    ) internal {
+    function _safeTransferFrom(address token, address from, address to, uint256 value) internal {
         if (value != 0) {
             require(token.code.length != 0);
-            (bool success, bytes memory data) = token.call(
-                abi.encodeWithSelector(
-                    IERC20Upgradeable.transferFrom.selector,
-                    from,
-                    to,
-                    value
-                )
-            );
+            (bool success, bytes memory data) =
+                token.call(abi.encodeWithSelector(IERC20Upgradeable.transferFrom.selector, from, to, value));
             require(success && (data.length == 0 || abi.decode(data, (bool))));
         }
     }
@@ -342,23 +279,15 @@ contract Voter is
     }
 
     // VE approval helpers
-    function _isAuthorized(
-        address owner,
-        address spender,
-        uint256 tokenId
-    ) private view returns (bool) {
-        return
-            spender != address(0) &&
-            (owner == spender ||
-                IERC721(_ve).isApprovedForAll(owner, spender) ||
-                IERC721(_ve).getApproved(tokenId) == spender);
+    function _isAuthorized(address owner, address spender, uint256 tokenId) private view returns (bool) {
+        return spender != address(0)
+            && (
+                owner == spender || IERC721(_ve).isApprovedForAll(owner, spender)
+                    || IERC721(_ve).getApproved(tokenId) == spender
+            );
     }
 
-    function _checkAuthorized(
-        address owner,
-        address spender,
-        uint256 tokenId
-    ) private view {
+    function _checkAuthorized(address owner, address spender, uint256 tokenId) private view {
         if (!_isAuthorized(owner, spender, tokenId)) {
             if (owner == address(0)) {
                 revert("ERC721: owner query for nonexistent token");
@@ -385,40 +314,27 @@ contract Voter is
         //create internal bribe contract to collect LP fees
         string memory _type = string.concat("Pearl LP Fees: ", _bribeType);
         _internal_bribe = IBribeFactory(bribefactory).createBribe(
-            _lzMainChainId,
-            _LzPoolChainId,
-            _pool,
-            owner(),
-            token0,
-            token1,
-            _type
+            _lzMainChainId, _LzPoolChainId, _pool, owner(), token0, token1, _type
         );
         isBribe[_internal_bribe] = true;
 
         //create external bribe
         _type = string.concat("Pearl Bribes: ", _bribeType);
         _external_bribe = IBribeFactory(bribefactory).createBribe(
-            _lzMainChainId,
-            _LzPoolChainId,
-            _pool,
-            owner(),
-            token0,
-            token1,
-            _type
+            _lzMainChainId, _LzPoolChainId, _pool, owner(), token0, token1, _type
         );
         isBribe[_external_bribe] = true;
 
-        (address _gauge, address _almGauge) = IGaugeV2Factory(gaugefactory)
-            .createGauge(
-                _lzMainChainId,
-                _LzPoolChainId,
-                factory,
-                _pool,
-                base, //rewardToken
-                address(this), //Distribution
-                _internal_bribe,
-                true
-            );
+        (address _gauge, address _almGauge) = IGaugeV2Factory(gaugefactory).createGauge(
+            _lzMainChainId,
+            _LzPoolChainId,
+            factory,
+            _pool,
+            base, //rewardToken
+            address(this), //Distribution
+            _internal_bribe,
+            true
+        );
 
         IERC20Upgradeable(base).approve(_gauge, type(uint256).max);
         internal_bribes[_gauge] = _internal_bribe;
@@ -431,13 +347,7 @@ contract Voter is
         supplyIndex[_gauge] = index;
         pools.push(_pool);
 
-        emit GaugeCreated(
-            _gauge,
-            msg.sender,
-            _internal_bribe,
-            _external_bribe,
-            _pool
-        );
+        emit GaugeCreated(_gauge, msg.sender, _internal_bribe, _external_bribe, _pool);
 
         return _gauge;
     }
@@ -453,10 +363,7 @@ contract Voter is
      * @param _adapterParams The adapter parameters required for Layerzero.
      * @return _gauge The address of the newly created gauge.
      */
-    function createGauge(
-        address _pool,
-        bytes memory _adapterParams
-    ) external payable returns (address _gauge) {
+    function createGauge(address _pool, bytes memory _adapterParams) external payable returns (address _gauge) {
         bool isPair = IPearlV2Factory(factory).isPair(_pool);
         require(isPair, "!pair");
 
@@ -467,18 +374,12 @@ contract Voter is
             // gov can create for any pool, even non-Pearl pairs
             require(isPair, "!_pool");
             // require(isWhitelisted[tokenA] && isWhitelisted[tokenB], "!whitelisted");
-            require(
-                ustb == address(0) || tokenA == ustb || tokenB == ustb,
-                "!ustb"
-            );
+            require(ustb == address(0) || tokenA == ustb || tokenB == ustb, "!ustb");
         }
 
         string memory _bribeType = string(
             abi.encodePacked(
-                "aMM-",
-                IERC20MetadataUpgradeable(tokenA).symbol(),
-                "/",
-                IERC20MetadataUpgradeable(tokenB).symbol()
+                "aMM-", IERC20MetadataUpgradeable(tokenA).symbol(), "/", IERC20MetadataUpgradeable(tokenB).symbol()
             )
         );
 
@@ -487,19 +388,10 @@ contract Voter is
 
         //notify main chain chain to create the child gauge for the pool
         if (!isMainChain) {
-            _adapterParams = _adapterParams.length > 2
-                ? _adapterParams
-                : defaultAdapterParams;
+            _adapterParams = _adapterParams.length > 2 ? _adapterParams : defaultAdapterParams;
             bytes memory _payload = abi.encode(_pool);
 
-            _lzSend(
-                lzMainChainId,
-                _payload,
-                payable(msg.sender),
-                address(0x0),
-                _adapterParams,
-                msg.value
-            );
+            _lzSend(lzMainChainId, _payload, payable(msg.sender), address(0x0), _adapterParams, msg.value);
         }
     }
 
@@ -528,17 +420,13 @@ contract Voter is
 
     function poke(address account) external nonReentrant isAllowed {
         require(msg.sender == _ve, "!ve");
-        uint256 lastVotedEpoch = (lastVoted[account] / EPOCH_DURATION) *
-            EPOCH_DURATION;
+        uint256 lastVotedEpoch = (lastVoted[account] / EPOCH_DURATION) * EPOCH_DURATION;
         uint256 currentEpoch = IMinter(minter).active_period();
         if (lastVotedEpoch < currentEpoch) return;
         _poke(account);
     }
 
-    function vote(
-        address[] memory _poolVote,
-        uint256[] memory _weights
-    ) external nonReentrant isAllowed {
+    function vote(address[] memory _poolVote, uint256[] memory _weights) external nonReentrant isAllowed {
         address account = msg.sender;
         require(_poolVote.length == _weights.length);
         lastVoted[account] = block.timestamp;
@@ -564,21 +452,14 @@ contract Voter is
         emit GaugeRevived(_gauge);
     }
 
-    function emitDeposit(
-        uint256 tokenId,
-        address account,
-        uint256 amount
-    ) external {
+    function emitDeposit(uint256 tokenId, address account, uint256 amount) external {
         require(isGauge[msg.sender]);
         require(isAlive[msg.sender]);
         emit Deposit(account, msg.sender, tokenId, amount);
     }
 
-    function claimBribes(
-        address[] memory _bribes,
-        address[][] memory _tokens
-    ) external isClaimAllowed {
-        for (uint256 i = _bribes.length; i != 0; ) {
+    function claimBribes(address[] memory _bribes, address[][] memory _tokens) external isClaimAllowed {
+        for (uint256 i = _bribes.length; i != 0;) {
             unchecked {
                 --i;
             }
@@ -586,11 +467,8 @@ contract Voter is
         }
     }
 
-    function claimFees(
-        address[] memory _fees,
-        address[][] memory _tokens
-    ) external isClaimAllowed {
-        for (uint256 i = _fees.length; i != 0; ) {
+    function claimFees(address[] memory _fees, address[][] memory _tokens) external isClaimAllowed {
+        for (uint256 i = _fees.length; i != 0;) {
             unchecked {
                 --i;
             }
@@ -599,7 +477,7 @@ contract Voter is
     }
 
     function distributeFees(address[] memory _gauges) external {
-        for (uint256 i = _gauges.length; i != 0; ) {
+        for (uint256 i = _gauges.length; i != 0;) {
             unchecked {
                 --i;
             }
@@ -632,7 +510,7 @@ contract Voter is
     }
 
     function distribute(uint256 start, uint256 finish) public {
-        for (uint256 x = start; x < finish; ) {
+        for (uint256 x = start; x < finish;) {
             distribute(gauges[pools[x]]);
             unchecked {
                 ++x;
@@ -641,7 +519,7 @@ contract Voter is
     }
 
     function distribute(address[] memory _gauges) external {
-        for (uint256 x = _gauges.length; x != 0; ) {
+        for (uint256 x = _gauges.length; x != 0;) {
             unchecked {
                 --x;
             }
@@ -649,11 +527,7 @@ contract Voter is
         }
     }
 
-    function emitWithdraw(
-        uint256 tokenId,
-        address account,
-        uint256 amount
-    ) external {
+    function emitWithdraw(uint256 tokenId, address account, uint256 amount) external {
         require(isGauge[msg.sender]);
         emit Withdraw(account, msg.sender, tokenId, amount);
     }
@@ -695,11 +569,7 @@ contract Voter is
         ustb = _ustb;
     }
 
-    function setNewBribe(
-        address _gauge,
-        address _internal,
-        address _external
-    ) external {
+    function setNewBribe(address _gauge, address _internal, address _external) external {
         require(msg.sender == emergencyCouncil);
         require(isGauge[_gauge] = true);
         internal_bribes[_gauge] = _internal;
@@ -724,14 +594,12 @@ contract Voter is
         factory = _factory;
     }
 
-    function setEpochController(
-        address _epochControllerAddress
-    ) external onlyGovernor {
+    function setEpochController(address _epochControllerAddress) external onlyGovernor {
         epochController = _epochControllerAddress;
     }
 
     function whitelist(address[] memory _token) public onlyGovernor {
-        for (uint256 i = _token.length; i != 0; ) {
+        for (uint256 i = _token.length; i != 0;) {
             unchecked {
                 --i;
             }
@@ -760,32 +628,21 @@ contract Voter is
     }
 
     function hasVoted(address _account) external view returns (bool) {
-        return
-            voted[_account] &&
-            lastVoted[_account] >= IMinter(minter).active_period();
+        return voted[_account] && lastVoted[_account] >= IMinter(minter).active_period();
     }
 
-    function estimateSendFee(
-        address _pool,
-        bool _useZro,
-        bytes memory _adapterParams
-    ) public view returns (uint nativeFee, uint zroFee) {
+    function estimateSendFee(address _pool, bool _useZro, bytes memory _adapterParams)
+        public
+        view
+        returns (uint256 nativeFee, uint256 zroFee)
+    {
         bytes memory _payload = abi.encode(_pool);
         uint16 _dstChainId = uint16(lzMainChainId);
         bytes memory trustedRemote = getTrustedRemote(_dstChainId);
-        (address _dstAddress, ) = abi.decode(trustedRemote, (address, address));
+        (address _dstAddress,) = abi.decode(trustedRemote, (address, address));
 
-        _adapterParams = _adapterParams.length > 2
-            ? _adapterParams
-            : defaultAdapterParams;
+        _adapterParams = _adapterParams.length > 2 ? _adapterParams : defaultAdapterParams;
 
-        return
-            lzEndpoint.estimateFees(
-                _dstChainId,
-                _dstAddress,
-                _payload,
-                _useZro,
-                _adapterParams
-            );
+        return lzEndpoint.estimateFees(_dstChainId, _dstAddress, _payload, _useZro, _adapterParams);
     }
 }
